refactor(Accounts): split AccountUsageDetails component [YTFRONT-5361]
