+++ packages/ui/src/ui/pages/accounts/tabs/detailed-usage/AccountUsageDetails/utils.ts
@@ -1,0 +1,3 @@
+import cn from 'bem-cn-lite';
+
+export const block = cn('account-usage-details');
